refact: combine DataSource, add lastRecognizedIdentity and resolveLastRecognizedIdentity
