+++ src/social-network/ui.ts
@@ -4,10 +4,15 @@ import { Person } from '../database/helpers/person'
 export interface SocialNetworkUI extends SocialNetworkWorkerAndUI, SocialNetworkUIDataSources {
     /** Should this UI content script activate? */
     shouldActivate(): boolean
+    resolveLastRecognizedIdentity(): void
 }
 export interface SocialNetworkUIDataSources {
     friendsRef: ValueRef<Person[]>
     myIdentitiesRef: ValueRef<Person[]>
+    /**
+     * The account that user is using (may not in the database)
+     */
+    lastRecognizedIdentity: ValueRef<Pick<Person, 'identifier' | 'nickname' | 'avatar'>>
 }
 
 const definedSocialNetworkUIs = new Set<SocialNetworkUI>()
@@ -18,7 +23,9 @@ export function activateSocialNetworkUI() {
         if (ui.shouldActivate()) {
             console.log('Activating UI provider', ui.networkIdentifier, ui)
             ui.init(env, {})
+            ui.resolveLastRecognizedIdentity()
             activatedSocialNetworkUI = ui
+            return
         }
 }
 export function defineSocialNetworkUI<T extends SocialNetworkUI>(UI: T) {
